app/requirements.txt
langchain_community
langchain_openai
elasticsearch
pandas
es_pandas
python-dotenv
